Migrated CamsJwt model to common
[skip ci]

Jira ticket: CAMS-418
