+++ backend/functions/lib/adapters/gateways/user-session-gateway.test.ts
@@ -1,6 +1,5 @@
 import { ConflictError, isConflictError, UserSessionGateway } from './user-session-gateway';
 import OktaGateway from './okta/okta-gateway';
-import { CamsJwtHeader } from '../types/authorization';
 import { ApplicationContext } from '../types/basic';
 import { createMockApplicationContext } from '../../testing/testing-utilities';
 import {
@@ -16,6 +15,7 @@ import { CamsSession } from '../../../../../common/src/cams/session';
 import { CamsRole } from '../../../../../common/src/cams/roles';
 import { urlRegex } from '../../../../../user-interface/src/lib/testing/testing-utilities';
 import { OFFICES } from '../../../../../common/src/cams/test-utilities/offices.mock';
+import { CamsJwtHeader } from '../../../../../common/src/cams/jwt';
 
 describe('user-session.gateway test', () => {
   const jwt = MockData.getJwt();
