+++ common/src/cams/jwt.ts
@@ -1,0 +1,21 @@
+export type CamsJwtHeader = {
+  typ: string;
+  [key: string]: unknown;
+};
+
+export type CamsJwtClaims = {
+  iss: string;
+  sub: string;
+  aud: string | string[];
+  exp: number;
+  nbf?: number;
+  iat?: number;
+  jti?: string;
+  groups: string[];
+  [key: string]: unknown;
+};
+
+export type CamsJwt = {
+  claims: CamsJwtClaims;
+  header: CamsJwtHeader;
+};
